Add aco_mem_new macro
Two types of allocation were done in the code, the former is the allocation
to predefined size and the latter callocation of single object.
I added a new macro aco_mem_calloc to handle the second case. I used the
calloc insted of malloc + memset for brevity.

This commit "breaks" C++ support.